Adding Microservices and Tenants configuration viewing
